major update
removed eletric sf compressor
removed cc gem compressor
removed cc gem decompressor
removed cc xtreme carbon press
(only removed them altogether all at once as very few (if more then one) have downloaded the addon yet and theyre endgame machines and not likely to have been crafted already (items will not be removed like this again in the future)
added gem recipes to regular machines
some refactoring
