--- src/main/java/me/jasperchasetoq/compressioncraft/implementation/machines/CcElectricSfCompressor.java
@@ -1,115 +1,0 @@
-package me.jasperchasetoq.compressioncraft.implementation.machines;
-
-import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
-import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
-import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
-import io.github.thebusybiscuit.slimefun4.core.attributes.RecipeDisplayItem;
-import io.github.thebusybiscuit.slimefun4.core.handlers.BlockBreakHandler;
-import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.MachineRecipe;
-import me.mrCookieSlime.Slimefun.api.BlockStorage;
-import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
-import org.bukkit.Material;
-import org.bukkit.block.Block;
-import org.bukkit.event.block.BlockBreakEvent;
-import org.bukkit.inventory.ItemStack;
-
-import javax.annotation.Nonnull;
-import javax.annotation.ParametersAreNonnullByDefault;
-import java.util.ArrayList;
-import java.util.List;
-
-public class CcElectricSfCompressor extends AContainer implements RecipeDisplayItem {
-
-
-    @ParametersAreNonnullByDefault
-    public CcElectricSfCompressor(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
-        super(itemGroup, item, recipeType, recipe);
-    }
-
-    @Nonnull
-    @Override
-    public List<ItemStack> getDisplayRecipes() {
-        List<ItemStack> displayRecipes = new ArrayList<>(recipes.size() * 4);
-
-        for (MachineRecipe recipe : recipes) {
-            displayRecipes.add(recipe.getInput()[0]);
-            displayRecipes.add(recipe.getOutput()[0]);
-        }
-
-        return displayRecipes;
-    }
-
-
-
-    @Override
-    protected void registerDefaultRecipes() {
-        registerRecipe(16, new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.STONE_CHUNK, 4)},
-                new ItemStack[]{new ItemStack(Material.COBBLESTONE, 1)});
-        registerRecipe(16, new ItemStack[]{new ItemStack(Material.FLINT, 8)},
-                        new ItemStack[]{new ItemStack(Material.COBBLESTONE, 1)});
-        registerRecipe(16, new ItemStack[]{new ItemStack(Material.CHARCOAL, 4)},
-                                new ItemStack[]{new ItemStack(Material.COAL, 1)});
-        registerRecipe(16, new ItemStack[]{new ItemStack(Material.COAL, 8)},
-                new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.CARBON, 1)});
-        registerRecipe(16, new ItemStack[]{new ItemStack(Material.COAL_BLOCK, 8)},
-                new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.CARBON, 9)});
-        registerRecipe(16, new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.CARBON, 4)},
-                new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.COMPRESSED_CARBON, 1)});
-        registerRecipe(16, new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.STEEL_INGOT, 8)},
-                new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.STEEL_PLATE, 1)});
-        registerRecipe(16, new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.REINFORCED_ALLOY_INGOT, 8)},
-                new SlimefunItemStack[]{new SlimefunItemStack(SlimefunItems.REINFORCED_PLATE, 1)});
-
-    }
-
-    @Override
-    public ItemStack getProgressBar() {
-        return new ItemStack(SlimefunItems.ELECTRIC_PRESS);
-    }
-
-    @Override
-    public String getInventoryTitle() {
-        return "&cElectric Slimefun Compressor";
-    }
-
-    @Override
-    public String getMachineIdentifier() {
-        return "CC_ELECTRIC_SF_COMPRESSOR";
-    }
-
-    @Override
-    public int getCapacity() {
-        return 1500;
-    }
-
-    @Override
-    public int getEnergyConsumption() {
-        return 150;
-    }
-
-    @Override
-    public int getSpeed() {
-        return 1;
-    }
-
-    public BlockBreakHandler onBreak() {
-        return new BlockBreakHandler(false, false) {
-
-            @Override
-            public void onPlayerBreak(BlockBreakEvent e, ItemStack item, List<ItemStack> drops) {
-                Block b = e.getBlock();
-                BlockMenu inv = BlockStorage.getInventory(b);
-
-                if (inv != null) {
-                    inv.dropItems(b.getLocation(), getInputSlots());
-                    inv.dropItems(b.getLocation(), getOutputSlots());
-                }
-            }
-        };
-    }
-
-}
-
-
