+++ src/main/java/me/jasperchasetoq/compressioncraft/implementation/machines/compressors/EndgameCcCompressor.java
@@ -1,4 +1,5 @@
-package me.jasperchasetoq.compressioncraft.implementation.machines.compressors.main;
+package me.jasperchasetoq.compressioncraft.implementation.machines.compressors;
+
 
 import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
@@ -19,10 +20,10 @@ import org.bukkit.inventory.ItemStack;
 import java.util.ArrayList;
 import java.util.List;
 
-public class AdvancedCcCompressor extends AContainer implements RecipeDisplayItem {
+public class EndgameCcCompressor extends AContainer implements RecipeDisplayItem {
 
 
-    public AdvancedCcCompressor(ItemGroup category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
+    public EndgameCcCompressor(ItemGroup category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
         addItemHandler(onBreak());
     }
@@ -35,14 +36,28 @@ public class AdvancedCcCompressor extends AContainer implements RecipeDisplayIte
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.GRAVEL, 8)}, new ItemStack[] {CompressionCraftItems.CC_GRAVEL_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.CLAY, 8)}, new ItemStack[] {CompressionCraftItems.CC_CLAY_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.COAL, 8)}, new ItemStack[] {CompressionCraftItems.CC_COAL_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.COAL_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_COAL_BLOCK_1});
         registerRecipe(4, new ItemStack[] {new ItemStack(Material.IRON_NUGGET, 8)}, new ItemStack[] {CompressionCraftItems.CC_IRON_NUGGET_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.IRON_INGOT, 8)}, new ItemStack[] {CompressionCraftItems.CC_IRON_INGOT_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.IRON_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_IRON_BLOCk_1});
         registerRecipe(4, new ItemStack[] {new ItemStack(Material.GOLD_NUGGET, 8)}, new ItemStack[]{CompressionCraftItems.CC_GOLD_NUGGET_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.GOLD_INGOT, 8)}, new ItemStack[] {CompressionCraftItems.CC_GOLD_INGOT_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.GOLD_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_GOLD_BLOCK_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.REDSTONE, 8)}, new ItemStack[] {CompressionCraftItems.CC_REDSTONE_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.REDSTONE_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_REDSTONE_BLOCK_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.LAPIS_LAZULI, 8)}, new ItemStack[] {CompressionCraftItems.CC_LAPIS_LAZULI_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.LAPIS_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_LAPIS_BLOCK_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.DIAMOND, 8)}, new ItemStack[] {CompressionCraftItems.CC_DIAMOND_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.DIAMOND_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_DIAMOND_BLOCK_1});
         registerRecipe(8, new ItemStack[] {new ItemStack(Material.EMERALD, 8)}, new ItemStack[] {CompressionCraftItems.CC_EMERALD_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.EMERALD_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_EMERALD_BLOCK_1});
+        registerRecipe(4, new ItemStack[] {new ItemStack(Material.AMETHYST_SHARD, 8)}, new ItemStack[] {CompressionCraftItems.CC_AMETHYST_SHARD_1});
+        registerRecipe(8, new ItemStack[] {new ItemStack(Material.AMETHYST_BLOCK, 8)}, new ItemStack[] {CompressionCraftItems.CC_AMETHYST_BLOCK_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.SMALL_AMETHYST_BUD, 8)}, new ItemStack[] {CompressionCraftItems.CC_SMALL_AMETHYST_BUD_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.MEDIUM_AMETHYST_BUD, 8)}, new ItemStack[] {CompressionCraftItems.CC_MEDIUM_AMETHYST_BUD_1});
+        registerRecipe(16, new ItemStack[] {new ItemStack(Material.LARGE_AMETHYST_BUD, 8)}, new ItemStack[] {CompressionCraftItems.CC_LARGE_AMETHYST_BUD_1});
+        registerRecipe(12, new ItemStack[] {new ItemStack(Material.AMETHYST_CLUSTER, 8)}, new ItemStack[] {CompressionCraftItems.CC_AMETHYST_CLUSTER_1});
+
     }
 
     @Override
@@ -62,20 +77,23 @@ public class AdvancedCcCompressor extends AContainer implements RecipeDisplayIte
         return new ItemStack(SlimefunItems.ELECTRIC_PRESS);
     }
     @Override
-    public String getInventoryTitle() {return "&6Advanced CC Compressor";}
+    public String getInventoryTitle() {return "&4End-Game CC Compressor";}
     @Override
     public String getMachineIdentifier() {
-        return "CC_COMPRESSOR_5";
+        return "CC_COMPRESSOR_6";
     }
     @Override
     public int getCapacity() {
-        return 1250;
+        return 1500;
     }
     @Override
     public int getEnergyConsumption() {
-        return 125;
+        return 150;
     }
-    @Override public int getSpeed() {return 5;}
+    @Override
+    public int getSpeed() {
+        return 6;
+    }
 
     public BlockBreakHandler onBreak() {
         return new BlockBreakHandler(false, false) {
